finito

+++ Systems Programming/ass4-electric-bugaloo/BankServer.c
@@ -9,29 +9,6 @@
 
 #include "Bank.h"
 
-/* temp
-#define KEY 666 // key used to access bank in shared mem
-#define MAXACCOUNTS 20 // max # of accounts in bank
-// Name for bank Semaphore
-#define BANKLOCK "bank"
-#include <stdlib.h>
-#include <stdio.h>
-#include <strings.h>
-#include <string.h>
-#include <unistd.h>
-#include  <sys/types.h>
-#include <sys/wait.h>
-#include <sys/ipc.h>
-#include <sys/shm.h>
-#include <fcntl.h>
-#include <sys/stat.h>
-#include <semaphore.h>
-#include <sys/socket.h>
-#include <netinet/in.h>
-#include <sys/prctl.h>
-#include <linux/limits.h>
-// end temp*/
-
 void endMethod(int signal);
 void suicide(int signal);
 void clientHandler(int clientSocket);
@@ -42,9 +19,11 @@ void getBalance(int clientSocket, int *in_session);
 void debitAccount(int clientSocket, char *commandArg, int *in_session);
 void creditAccount(int clientSocket, char *commandArg, int *in_session);
 
-int spawnedProcesses;
+int spawnedProcesses; // number of processes spawned
 int printBankPID; // pid of bank printing process
-Bank *bank;
+Bank *bank; // ptr to bank
+int bankID; // shmid for bank
+int originalParentPPID;
 
 int main(int argc, char *argv[])
 {
@@ -54,22 +33,19 @@ int main(int argc, char *argv[])
     }
 
     spawnedProcesses = 0;
+    originalParentPPID = getppid();
 
     signal(SIGINT, endMethod); // set function to run at program termination
 
-    /* Init bank and put into shared memory */
-
     // allocate space for bank in shared memory
-    int bankID = shmget(KEY, sizeof(Bank), IPC_CREAT | 0666);
+    bankID = shmget(KEY, sizeof(Bank), IPC_CREAT | 0666);
 
     bank = (Bank*) shmat(bankID, NULL, 0); // get bank from shared mem
     createBank(bank); // init bank
 
-    /* Init bank semaphores (locks) */
-
+    // make account locks
     char semaphoreName[100];
     int i;
-    // make account locks
     for(i = 0; i < MAXACCOUNTS; i++)
     {
         bzero((void*) semaphoreName, 100); // zero out buffer
@@ -77,18 +53,18 @@ int main(int argc, char *argv[])
         sem_unlink(semaphoreName); // in case it already exists
         sem_open(semaphoreName, O_CREAT, 0644, 1); // create account lock
     }
+
     sem_unlink(BANKLOCK); // in case it already exists
     sem_open(BANKLOCK, O_CREAT, 0644, 1); // create bank lock
 
-    /* Spawn process to print bank contents on an interval*/
-
-    char *args[] = {"PrintBank", 0}; // args to be passed to child process
-    char *env[] = { 0 }; // environment to be passed to child process
-
-    // get path name
+    // args and env to be passed to child process
+    char *args[] = {"PrintBank", 0};
+    char *env[] = { 0 };
+    // get path name for exe to be run in new process
     char filePath[PATH_MAX];
     realpath("PrintBank", filePath);
 
+    // spawn bank-printing process
     spawnedProcesses++;
     pid_t pid = fork();
     printBankPID = pid;
@@ -106,13 +82,10 @@ int main(int argc, char *argv[])
         exit(EXIT_FAILURE);
     }
 
-    /* Create socket and bind to specified server port */
-
     int port = atoi(argv[1]); // port number to bind socket to
-
     // build socket
-    int serverSocket = socket(AF_INET, SOCK_STREAM, 0);
-    if(serverSocket < 0){
+    int serverSocket;
+    if((serverSocket = socket(AF_INET, SOCK_STREAM, 0)) < 0){
         printf("%s", "Error opening socket.");
         exit(EXIT_FAILURE);
     }
@@ -132,10 +105,7 @@ int main(int argc, char *argv[])
         exit(EXIT_FAILURE);
     }
 
-    // set socket to listen for client connects
     listen(serverSocket, 10);
-
-    // listen and accept incoming requests
     int clientSocket; // file descriptor for client socket
     struct sockaddr_in clientAddressInfo;
     socklen_t len = sizeof(clientAddressInfo);
@@ -162,9 +132,10 @@ int main(int argc, char *argv[])
             clientHandler(clientSocket); // run client handler
         }
         else if(pid < 0){ // fork failed
-            printf("%s", "Error spawning clientHandler process");
+            printf("%s", "Error spawning clientHandler process\n");
             exit(EXIT_FAILURE);
         }
+        printf("%s", "Recieved a client connection.\n");
         spawnedProcesses++;
     }
 
@@ -175,31 +146,23 @@ void clientHandler(int clientSocket)
 {
     // connection error
     if(clientSocket < 0){
-        printf("%s", "Error connecting to clientSocket.");
+        printf("%s", "Error connecting to clientSocket.\n");
         exit(EXIT_FAILURE);
     }
 
     int in_session = -1;
 
-    // create buffer for socket messages
-    char buffer[256];
-    bzero((void*) buffer, 256);
-
     // read from client
+    char buffer[256]; // buffer for socket messages
     while(1){
-        if((read(clientSocket, buffer, 255)) <= 0){
-            printf("%s", "No client message or error, not sure lol");
+        bzero((void*) buffer, 256);
+        if((read(clientSocket, buffer, 255)) <= 0){ // no response
             exit(EXIT_SUCCESS);
         }
-
-        // display message for debug purposed
-        printf("Client message: %s", buffer);
 
         // parse command
         char *command = strtok(buffer, " \n");
         char *commandArg = strtok(NULL, " \n");
-
-        /* handle client operation */
 
         if(strcmp(command, "open") == 0){ // open account
            openAccount(clientSocket, commandArg, &in_session);
@@ -207,11 +170,11 @@ void clientHandler(int clientSocket)
         else if(strcmp(command, "start") == 0){ // start account session
             startAccount(clientSocket, commandArg, &in_session);
         }
-        else if(buffer[0] == 'c'){ // credit account
-
+        else if(strcmp(command, "credit") == 0){ // credit account
+            creditAccount(clientSocket, commandArg, &in_session);
         }
-        else if(buffer[0] == 'd'){ // debit account
-
+        else if(strcmp(command, "debit") == 0){ // debit account
+            debitAccount(clientSocket, commandArg, &in_session);
         }
         else if(strcmp(command, "balance") == 0){ // get balance
             getBalance(clientSocket, &in_session);
@@ -225,6 +188,7 @@ void clientHandler(int clientSocket)
         }
     }
 
+    return;
 }
 
 void openAccount(int clientSocket, char *commandArg, int *in_session)
@@ -242,7 +206,7 @@ void openAccount(int clientSocket, char *commandArg, int *in_session)
     int exists = 0;
     for(i = 0; i < MAXACCOUNTS; i++)
     {
-        if(*bank->accounts[i].name){ // taken
+        if(bank->accounts[i].name[0] != '\0'){ // taken
             // check if desired account to open already exists
             exists = (exists == 0) ? (strcmp(bank->accounts[i].name, commandArg) == 0) : 1;
             continue;
@@ -252,7 +216,10 @@ void openAccount(int clientSocket, char *commandArg, int *in_session)
         }
     }
 
-    if(i < MAXACCOUNTS) // found free spot
+    if(exists){
+        write(clientSocket, "Account already exists", 22);
+    }
+    else if(i < MAXACCOUNTS)
     {
         strcpy(bank->accounts[i].name, commandArg);
         write(clientSocket, "Account created!", 16);
@@ -282,7 +249,7 @@ void startAccount(int clientSocket, char *commandArg, int *in_session)
     int i;
     for(i = 0; i < MAXACCOUNTS; i++)
     {
-        if(bank->accounts[i].name){
+        if(bank->accounts[i].name[0] == '\0'){ // free account; end of list
             break;
         }
 
@@ -298,10 +265,13 @@ void startAccount(int clientSocket, char *commandArg, int *in_session)
             bank->accounts[i].in_use = 1;
             *in_session = i;
             write(clientSocket, "Connected successfully!", 23);
+            break;
         }
     }
 
-    write(clientSocket, "Error: bank with specified account name not found.", 50); // bank not found
+    if(i >= MAXACCOUNTS){ // bank not found
+        write(clientSocket, "Error: bank with specified account name not found.", 50);
+    }
 
     sem_post(bankLock);
 
@@ -313,6 +283,7 @@ void finishSession(int clientSocket, int *in_session)
     // end the account session if they are in one
     if(*in_session >= 0){
         bank->accounts[*in_session].in_use = 0;
+        *in_session = -1;
     }
 
     write(clientSocket, "Successful exit!", 16);
@@ -348,7 +319,7 @@ void getBalance(int clientSocket, int *in_session)
 void debitAccount(int clientSocket, char *commandArg, int *in_session)
 {
     if(*in_session < 0){
-        write(clientSocket, "Can't credit account, not in session", 36);
+        write(clientSocket, "Can't debit account, not in session", 36);
         return;
     }
 
@@ -395,24 +366,23 @@ void creditAccount(int clientSocket, char *commandArg, int *in_session)
 
     sem_post(accountLock);
 
+    write(clientSocket, "Account credited!", 17);
+
     return;
 }
 
 void endMethod(int signal)
 {
-    // kill bankPrinting process
-    kill(printBankPID, SIGKILL);
-
-    // wait on all spawned processes
-    int status;
-    int i;
-    for(i = 0; i < spawnedProcesses; i++)
-    {
-        wait(&status);
-        if(status){
-            // should do somthing along these lines
-        }
+    if(getppid() != originalParentPPID){ // not parent
+        exit(EXIT_SUCCESS);
     }
+
+    kill(printBankPID, SIGKILL); // kill bankPrinting process
+
+    // remove shared memory
+    shmdt((void*) bank);
+    shmctl(bankID, IPC_RMID, NULL);
+    printf("%s\n", "\nRemoved Shared Memory\n");
 
     exit(EXIT_SUCCESS);
 }
